+++ timezones.nim
@@ -1,49 +1,35 @@
 ##[
-Examples:
+    .. code-block :: nim
+        import times
+        import timezones
 
-.. code-block:: nim
-    import times
-    import timezones
+        # Create a timezone representing a static offset from UTC.
+        let zone = tz"+02:30"
+        echo initDateTime(1, mJan, 2000, 12, 00, 00, zone)
+        # => 2000-01-01T12:00:00+02:30
 
-    # Create a timezone representing a static offset from UTC.
-    let zone = tz"+02:30"
-    echo initDateTime(1, mJan, 2000, 12, 00, 00, zone)
-    # => 2000-01-01T12:00:00+02:30
+        # Static offset timezones can also be created with the proc ``staticTz``,
+        # which is preferable if the offset is only known at runtime.
+        doAssert zone == staticTz(hours = -2, minutes = -30)
 
-    # Static offset timezones can also be created with the proc ``staticTz``,
-    # which is preferable if the offset is only known at runtime.
-    doAsert zone == staticTz(hours = -2, minutes = -30)
+        # Create a timezone representing a timezone in the IANA timezone database.
+        let stockholm = tz"Europe/Stockholm"
+        echo initDateTime(1, mJan, 1850, 00, 00, 00, stockholm)
+        # => 1850-01-01T00:00:00+01:12
 
-    # Create a timezone representing a timezone in the IANA timezone database.
-    let stockholm = tz"Europe/Stockholm"
-    echo initDateTime(1, mJan, 1850, 00, 00, 00, stockholm)
-    # => 1850-01-01T00:00:00+01:12
+        # Like above, but returns a `TimezoneInfo` object which contains some
+        # extra metadata.
+        let stockholmInfo = tzInfo"Europe/Stockholm"
+        # Countries are specified with it's two character country code,
+        # see ISO 3166-1 alpha-2.
+        doAssert stockholmInfo.countries == @["SE"]
+        doAssert stockholmInfo.timezone == stockholm
 
-    # Get a list of timezones used by a country.
-    # The country is specified with it's two character country code,
-    # see ISO 3166-1 alpha-2.
-    let sweden = tzNames("SE")
-    echo sweden
-    # => @["Europe/Stockholm"]
-
-    # Note that some countries use many different timezones.
-    let usa = tzNames("US")
-    echo usa
-    # => @[
-    #   "America/New_York",    "America/Adak",      "America/Phoenix",
-    #   "America/Yakutat",     "Pacific/Honolulu",  "America/Nome",
-    #   "America/Los_Angeles", "America/Detroit",   "America/Chicago",
-    #   "America/Boise",       "America/Juneau",    "America/Metlakatla",
-    #   "America/Anchorage",   "America/Menominee", "America/Sitka",
-    #   "America/Denver"
-    # ]
-
-    # Get a list of countries that are known to use a timezone.
-    # Note that some timezones are used by multiple countries.
-    let bangkok = tz"Asia/Bangkok"
-    echo bangkok.countries
-    # => @["TH", "KH", "LA", "VN"]
+        # Note that some timezones are used by multiple countries.
+        let bangkok = tzInfo"Asia/Bangkok"
+        doAssert bangkok.countries == @["TH", "KH", "LA", "VN"]
 ]##
+
 
 #TODO:
 #[
@@ -59,46 +45,47 @@ when not defined(js):
 when not defined(nimdoc):
     export coordinates
 
+when defined(timezonesPath) and defined(timezonesNoEmbeed):
+    {.warning: "Both `timezonesPath` and `timezonesNoEmbeed` was passed".}
+
 type
-    Country* = string ## A country is represented by
-        ## a two character country code, using ISO 3166-1 alpha-2.
+    Country* = string ## \
+        ## A country is represented by a two character country code,
+        ## using ISO 3166-1 alpha-2.
         ##
         ## See https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2.
 
     # Instead of re-exporting timezonedbs.TzData we need to do it like this
     # to avoid having the fields of timezonedbs.TzData re-exported as well.
     TimezoneDbImpl = distinct timezonedbs.TimezoneDb
-    TimezoneDb* = TimezoneDbImpl ## Contains the timezone data
-                         ## for a specific IANA timezone database version.
+    TimezoneDb* = TimezoneDbImpl ## \
+        ## A collection of loaded timezones, typically from a specific release
+        ## of the IANA timezone database.
 
-    TimezoneInfo* = object ## A timezone with additional meta data attached.
-        timezone*: Timezone             ## The timezone as a ``times.Timezone``
-                                        ## object.
-        countries*: seq[Country]        ## Get a list of countries that are
-                                        ## known to use this timezone.
-                                        ## Note that some countries use
-                                        ## multiple timezones.
-        location*: Option[Coordinates]  ## Get the coordinates of a timezone.
-                                        ## This is generally the coordinates
-                                        ## of the city in the timezone name.
-                                        ## E.g ``db.location"Europe/Stockholm"``
-                                        ## will give the the coordinates of
-                                        ## Stockholm, the capital of Sweden.
+    TimezoneInfo* = object ## \
+        ## A timezone with additional metadata attached.
+        timezone*: Timezone ## \
+            ## The timezone as a ``times.Timezone`` object.
+        countries*: seq[Country] ## \
+            ## Get a list of countries that are
+            ## known to use this timezone.
+            ## Note that some countries use
+            ## multiple timezones.
+        location*: Option[Coordinates]  ## \
+            ## Get the coordinates of a timezone. This is generally the
+            ## coordinates of the city in the timezone name.
+            ## E.g ``db.location"Europe/Stockholm"`` will give the the
+            ## coordinates of Stockholm, the capital of Sweden.
 
-proc version*(db: TimezoneDb): string =
-    ## The version of the IANA timezone database being represented by ``db``.
-    ## The string consist of the year plus a letter. For example, ``"2018a"``
-    ## is the first database release of 2018, ``"2018b"``
-    ## the second one and so on.
-    timezonedbs.TimezoneDb(db).version
+var defaultTzDb {.threadvar.}: TimezoneDb
 
-proc getTz(db: TimezoneDb,
-           tzName: string): (bool, TimezoneInternal) {.inline.} =
+proc getTz(db: TimezoneDb, tzName: string): (bool, TimezoneInternal)
+        {.inline, raises: [].} =
     let tz = timezonedbs.TimezoneDb(db).tzByName.getOrDefault(tzName)
     if tz.name != "":
         result = (true, tz)
 
-proc newTimezone(tzName: string, offset: int): Timezone =
+proc newTimezone(tzName: string, offset: int): Timezone {.raises: [].} =
     proc zoneInfoFromAdjTime(adjTime: Time): ZonedTime {.locks: 0.} =
         result.isDst = false
         result.utcOffset = offset
@@ -111,7 +98,7 @@ proc newTimezone(tzName: string, offset: int): Timezone =
 
     result = newTimezone(tzName, zoneInfoFromTime, zoneInfoFromAdjTime)
 
-proc tz*(db: TimezoneDb, tzName: string): Timezone {.inline.} =
+proc tz*(db: TimezoneDb, tzName: string): Timezone {.raises: [ValueError].} =
     ## Retrieve a timezone from a timezone name, where the timezone name is one
     ## of the following:
     ## - The string ``"LOCAL"``, representing the systems local timezone.
@@ -122,26 +109,30 @@ proc tz*(db: TimezoneDb, tzName: string): Timezone {.inline.} =
     ## - A timezone name from the
     ##   `IANA timezone database <https://www.iana.org/time-zones>`_.
     ##   See
-    ##   `wikipedia <https://en.wikipedia.org/wiki/List_of_tz_database_time_zones>`_.
-    ##   for a list of available timezone names.
+    ##   `wikipedia
+    ##   <https://en.wikipedia.org/wiki/List_of_tz_database_time_zones>`_
+    ##   for a list of available timezone names. Note that there is no guranteas
+    ##   about what timezones are available in ``db``, but as a special rule
+    ##   ``"Etc/UTC"`` is always supported.
     ##
     ## In case ``tzName`` does not follow any of these formats, or the timezone
-    ## name doesn't exist in the database, a ``ValueError`` exception is raised.
+    ## name doesn't exist in the database, a ``ValueError`` is raised.
     if tzName.len == 0:
         raise newException(ValueError, "Timezone name can't be empty")
     if tzName == "LOCAL":
         result = local()
+    elif tzName == "Etc/UTC":
+        result = utc()
     elif tzName[0] in {'-', '+'}:
         template error =
             raise newException(ValueError,
                 "Invalid static timezone offset: " & tzName)
         template parseTwoDigits(str: string, idx: int): int =
             if str[idx] notin {'0'..'9'} or str[idx + 1] notin {'0'..'9'}:
-                echo "err"
                 error()
-            (str[idx].ord - '0'.ord) * 10 + (str[idx].ord - '0'.ord)
+            (str[idx].ord - '0'.ord) * 10 + (str[idx + 1].ord - '0'.ord)
 
-        let sign = if tzName[0] == '-': -1 else: +1
+        let sign = if tzName[0] == '-': 1 else: -1
         case tzName.len
         of 6:
             if tzName[3] != ':':
@@ -149,7 +140,7 @@ proc tz*(db: TimezoneDb, tzName: string): Timezone {.inline.} =
             let h = parseTwoDigits(tzName, 1)
             let m = parseTwoDigits(tzName, 4)
             let offset = h * 3600 + m * 60
-            result = newTimezone(tzName, offset)
+            result = newTimezone(tzName, sign * offset)
         of 9:
             if tzName[3] != ':' or tzName[6] != ':':
                 error()
@@ -157,7 +148,7 @@ proc tz*(db: TimezoneDb, tzName: string): Timezone {.inline.} =
             let m = parseTwoDigits(tzName, 4)
             let s = parseTwoDigits(tzName, 7)
             let offset = h * 3600 + m * 60 + s
-            result = newTimezone(tzName, offset)
+            result = newTimezone(tzName, sign * offset)
         else:
             error()
     else:
@@ -167,15 +158,30 @@ proc tz*(db: TimezoneDb, tzName: string): Timezone {.inline.} =
                 "Timezone does not exist in database: " & tzName)
         result = newTimezone(tz)
 
-proc tzInfo*(db: TimezoneDb, tzName: string): TimezoneInfo {.inline.} =
-    ## Retrieve a timezone with additional meta data.
-    ## Supports the same formats as ``db.tz(...)``.
+proc tz*(tzName: string): Timezone {.inline, raises: [ValueError].} =
+    ## Convenience proc using the default timezone database.
+    runnableExamples:
+        import times
+        let stockholm = tz"Europe/Stockholm"
+        let dt = initDateTime(1, mJan, 1850, 00, 00, 00, stockholm)
+        doAssert $dt == "1850-01-01T00:00:00+01:12"
+    defaultTzDb.tz(tzName)
+
+proc tzInfo*(db: TimezoneDb, tzName: string): TimezoneInfo
+             {.raises: [ValueError].}=
+    ## Retrieve a timezone with additional metadata.
+    ##
+    ## The ``tzName`` parameter follows the same format as ``db.tz(...)``.
+    ##
+    ## In case ``tzName`` has an invalid format, or the timezone name doesn't
+    ## exist in the database, a ``ValueError`` is raised.
     if tzName.len == 0:
         raise newException(ValueError, "Timezone name can't be empty")
 
     if tzName == "LOCAL":
         result.timezone = local()
-        return
+    elif tzName == "Etc/UTC":
+        result.timezone = utc()
     elif tzName[0] in {'-', '+'}:
         result.timezone = db.tz(tzname)
     else:
@@ -187,69 +193,31 @@ proc tzInfo*(db: TimezoneDb, tzName: string): TimezoneInfo {.inline.} =
         result.countries = tz.countries.mapIt($it)
         result.location = tz.location
 
+proc tzInfo*(tzName: string): TimezoneInfo {.inline, raises: [ValueError].} =
+    ## Convenience proc using the default timezone database.
+    runnableExamples:
+        import times, options
+        let stockholmInfo = tzInfo"Europe/Stockholm"
+        doAssert stockholmInfo.timezone == tz"Europe/Stockholm"
+        doAssert stockholmInfo.countries == @["SE"]
+        doAssert $stockholmInfo.location == "Some(59° 20′ 0″ N 18° 3′ 0″ E)"
+    defaultTzDb.tzInfo(tzName)
 
-proc parseJsonTimezones*(content: string): TimezoneDb =
-    ## Parse a timezone database from its JSON representation.
-    parseTzData(content).TimezoneDb
+proc setDefaultTzDb*(db: TimezoneDb) =
+    ## Sets the timezone database that will be used for ``tz(tzName)`` and
+    ## ``tzInfo(tzName)``. The default timezone database is stored in a thread
+    ## local variable, so calling this only affects the calling thread!
+    defaultTzDb = db
 
-when not defined(js):
-    proc parseJsonTimezones*(s: Stream): TimezoneDb =
-        ## Parse a timezone database from its JSON representation.
-        parseTzData(s).TimezoneDb
+proc getDefaultTzDb*(): TimezoneDb =
+    ## Gets the timezone database that is used for ``tz(tzName)`` and
+    ## ``tzInfo(tzName)``. The default timezone database is stored in a thread
+    ## local varaible, so calling ``setDefaultTzDb`` only affects the calling
+    ## thread!
+    defaultTzDb
 
-    proc loadJsonTimezones*(path: string): TimezoneDb =
-        ## Load a timezone database from a JSON file.
-        let fs = openFileStream(path, fmRead)
-        defer: fs.close
-        parseTzData(fs).TimezoneDb
-
-# xxx the silly default path is because it's relative to "timezonedbs.nim"
-when not defined(nimsuggest):
-    when not defined(timezonesPath):
-        const timezonesPath = "./" & Version & ".json"
-    else:
-        const timezonesPath {.strdefine.} = ""
-        # isAbsolute isn't available for JS
-        when not defined(js):
-            when not timezonesPath.isAbsolute:
-                {.error: "Path to custom tz data file must be absolute: " &
-                    timezonesPath.}
-
-        {.hint: "Embedding custom tz data file: " & timezonesPath.}
-
-when defined(timezonesPath) and defined(timezonesNoEmbeed):
-    {.warning: "Both `timezonesPath` and `timezonesNoEmbeed` was passed".}
-
-when not defined(timezonesNoEmbeed) or defined(nimdoc):
-    const content = staticRead timezonesPath
-
-    let embeededTzDataImpl = parseTzData(content).TimezoneDb
-    let EmbeededTzData* = embeededTzDataImpl ## The embeeded tzdata.
-        ## Not available if ``-d:timezonesNoEmbeed`` is used.
-
-    when not defined(nimdoc):
-        template EmbeededTzdb*(): TimezoneDb
-            {.deprecated: "Renamed to EmbeededTzData".} = EmbeededTzData
-
-    proc tz*(tzName: string): Timezone =
-        ## Convenience proc using the embeeded timezone database.
-        runnableExamples:
-            import times
-            let stockholm = tz"Europe/Stockholm"
-            let dt = initDateTime(1, mJan, 1850, 00, 00, 00, stockholm)
-            doAssert $dt == "1850-01-01T00:00:00+01:12"
-        EmbeededTzData.tz(tzName)
-
-    proc tzInfo*(tzName: string): TimezoneInfo =
-        ## Convenience proc using the embeeded timezone database.
-        runnableExamples:
-            import times
-            let stockholm = tz"Europe/Stockholm"
-            let dt = initDateTime(1, mJan, 1850, 00, 00, 00, stockholm)
-            doAssert $dt == "1850-01-01T00:00:00+01:12"
-        EmbeededTzData.tzInfo(tzName)
-
-proc staticTz*(hours, minutes, seconds: int = 0): Timezone {.noSideEffect.} =
+proc staticTz*(hours, minutes, seconds: int = 0): Timezone
+               {.noSideEffect, raises: [].} =
     ## Create a timezone using a static offset from UTC.
     runnableExamples:
         import times
@@ -279,6 +247,51 @@ proc staticTz*(hours, minutes, seconds: int = 0): Timezone {.noSideEffect.} =
             "+" & offsetStr
 
     result = newTimezone(tzName, offset)
+
+proc parseTzDb*(content: string): TimezoneDb =
+    ## Parse a timezone database from its JSON representation.
+    parseTzData(content).TimezoneDb
+
+when not defined(js):
+    proc parseTzDb*(s: Stream): TimezoneDb =
+        ## Parse a timezone database from its JSON representation.
+        parseTzData(s).TimezoneDb
+
+    proc loadTzDb*(path: string): TimezoneDb =
+        ## Load a timezone database from a JSON file.
+        let fs = openFileStream(path, fmRead)
+        defer: fs.close()
+        parseTzData(fs).TimezoneDb
+
+when not defined(nimsuggest):
+    when not defined(timezonesPath):
+        const timezonesPath = "./" & Version & ".json"
+    else:
+        const timezonesPath {.strdefine.} = ""
+        # isAbsolute isn't available for JS
+        when not defined(js):
+            when not timezonesPath.isAbsolute:
+                {.error: "Path to custom tz data file must be absolute: " &
+                    timezonesPath.}
+
+        {.hint: "Embedding custom tz data file: " & timezonesPath.}
+
+when not defined(timezonesNoEmbeed) or defined(nimdoc):
+    const content = staticRead timezonesPath
+    defaultTzDb = parseTzData(content).TimezoneDb
+
+proc version*(db: TimezoneDb): string {.raises: [].} =
+    ## The version of the IANA timezone database being represented by ``db``.
+    ## The string consist of the year plus a letter. For example, ``"2018a"``
+    ## is the first database release of 2018, ``"2018b"``
+    ## the second one and so on.
+    ##
+    ## If the version is unknown, returns `""`.
+    timezonedbs.TimezoneDb(db).version
+
+proc tzNames*(db: TimezoneDb): seq[string] {.raises: [].} =
+    ## Retrieve a list of all available timezones in ``db``.
+    toSeq(timezonedbs.TimezoneDb(db).tzByName.keys)
 
 # Trick to simplify doc gen.
 # This might break in the future