+++ timezones/posixtimezones.nim
@@ -1,9 +1,27 @@
-# Fhis file is mostly a direct translation of PosixTimeZone from phobos,
-# the standard library of D.
+##[
+  The `posixtimezones` module implements handling of the system timezone
+  files on posix systems. It's not available on non-posix systems.
+
+  Usage
+  -----
+  .. code-block :: nim
+    import timezones, timezones/posixtimezones
+    # Load a timezone from the systems timezone dir
+    let zone1 = loadPosixTz("Europe/Stockholm")
+    # Load a timezone with an absolute path to the file
+    let zone2 = loadPosixTz("/path/to/timezone/file")
+    # Load all available timezones from the systems timezone dir
+    let db = loadPosixTzDb()
+    # Timezones can now be loaded from the db instead of from the file system
+    let zone3 = db.tz("Europe/Stockholm")
+]##
 
 import std / [times, options, sequtils, os, algorithm, sugar, strutils]
 import private / [coordinates, zone1970, timezonedbs]
 from .. / timezones import TimezoneInfo
+
+when not defined(posix):
+  {.error: "The `posixtimezones` module is only available on posix systems.".}
 
 type
   TTInfo = object
@@ -28,9 +46,10 @@ type
   LeapSecond = object
     time: int64  ## The time_t when the leap second occurs.
     total: int32 ## The total number of leap seconds to be applied after
-                  ## the corresponding leap second.
+                 ## the corresponding leap second.
 
-  TzFileParsingError* = object of ValueError
+  TzFileParsingError* = object of ValueError ## \
+      ## Exception which is raised when parsing a posix timezone file fails.
 
 const Zeros15 = @[byte 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0] # 15 zeros
 
@@ -38,14 +57,16 @@ template tzAssert(condition: untyped) =
   if not condition:
     raise newException(TzFileParsingError, "Invalid tz file")
 
-proc readVal[T: int8|int16|int32|int64|byte|bool|char](f: File): T =
+type Readable = int8|int16|int32|int64|byte|bool|char
+proc readVal[T: Readable](f: File, _: typedesc[T]): T
+    {.raises: [IOError, TzFileParsingError].} =
   # TODO: Investigate if this is the best implementation
   var buffer: array[T.sizeof, byte]
   tzAssert not endOfFile(f)
   tzAssert readBytes(f, buffer, 0, T.sizeof) == T.sizeof
   var conv = buffer
   # The TZ file is in big endian, so in case we're on little endian
-    # we must convert it.
+  # we must convert it.
   when cpuEndian == littleEndian:
     when T.sizeof == 2:
       conv[0] = buffer[1]
@@ -66,7 +87,8 @@ proc readVal[T: int8|int16|int32|int64|byte|bool|char](f: File): T =
       conv[7] = buffer[0]
   result = cast[T](conv)
 
-proc readArray[T](f: File, len: int): seq[T] =
+proc readArray(f: File, T: typedesc, len: int): seq[T]
+    {.raises: [IOError, TzFileParsingError].} =
   when T.sizeof != 1:
     raise newException(ValueError, "T must have sizeof 1")
   else:
@@ -76,12 +98,12 @@ proc readArray[T](f: File, len: int): seq[T] =
     result = cast[seq[T]](buffer)
 
 proc readStr(f: File, len: int): string =
-  result = cast[string](readArray[char](f, len))
+  result = cast[string](readArray(f, char, len))
 
 proc readTempTTInfo(f: File): TempTTInfo =
-  result.tt_gmtoff = readVal[int32](f)
-  result.tt_isdst = readVal[bool](f)
-  result.tt_abbrind = readVal[byte](f)
+  result.tt_gmtoff = readVal(f, int32)
+  result.tt_isdst = readVal(f, bool)
+  result.tt_abbrind = readVal(f, byte)
 
 proc extractString(chars: seq[char], start: byte): string =
   # TODO: optimize
@@ -90,55 +112,53 @@ proc extractString(chars: seq[char], start: byte): string =
     result.add chars[idx]
     idx.inc
 
-proc loadLocation(tzPath: string, dir: string):
-    Option[(Coordinates, seq[string])] {.raises: [TzFileParsingError].} =
+proc loadLocation(dir, path: string): Option[(Coordinates, seq[string])]
+                  {.raises: [TzFileParsingError].} =
   let tab = dir / "zone1970.tab"
   if not tab.fileExists:
     return
   try:
-    for countries, coords, tzName in zone1970Entries(tab, find = tzPath):
+    for countries, coords, tzName in zone1970Entries(tab, find = path):
       return some((coords, countries))
-  except Exception as e:
-    raise newException(TzFileParsingError,
-      "Failed to parse zone1970.tab", e)
+  except ValueError, IOError:
+    let e = getCurrentException()
+    raise newException(TzFileParsingError, "Failed to parse zone1970.tab", e)
 
-proc loadTzInternal(path, dir: string,
+proc loadTzInternal(dir, path: string,
                     loadLocation = false): TimezoneInternal
                     {.raises: [IOError, TzFileParsingError].} =
-  let fullPath =
-      if path.isAbsolute: path
-      else: dir / path
   let isGmtZone = path.startsWith("GMT")
 
-  let file = open(fullPath, fmRead)
+  let file = open(dir / path, fmRead)
+  defer: close(file)
   tzAssert readStr(file, 4) == "TZif"
-  let version = readVal[char](file)
+  let version = readVal(file, char)
   tzAssert version in {'\0', '2', '3'}
-  tzAssert readArray[byte](file, 15) == Zeros15
+  tzAssert readArray(file, byte, 15) == Zeros15
 
   # The number of UTC/local indicators stored in the file.
-  let tzh_ttisgmtcnt = readVal[int32](file)
+  let tzh_ttisgmtcnt = readVal(file, int32)
   # The number of standard/wall indicators stored in the file.
-  let tzh_ttisstdcnt = readVal[int32](file)
+  let tzh_ttisstdcnt = readVal(file, int32)
   # The number of leap seconds for which data is stored in the file.
-  let tzh_leapcnt = readVal[int32](file)
+  let tzh_leapcnt = readVal(file, int32)
   # The number of "transition times" for which data is stored in the file.
-  let tzh_timecnt = readVal[int32](file)
+  let tzh_timecnt = readVal(file, int32)
   # The number of "local time types" for which data is stored in the file(must not be zero).
-  let tzh_typecnt = readVal[int32](file)
+  let tzh_typecnt = readVal(file, int32)
   # The number of characters of "timezone abbreviation strings" stored in the file.
-  let tzh_charcnt = readVal[int32](file)
+  let tzh_charcnt = readVal(file, int32)
 
   # time_ts where DST transitions occur.
   var transitionTimeTs = newSeq[int64](tzh_timecnt)
   for idx in 0 ..< tzh_timecnt:
-    transitionTimeTs[idx] = readVal[int32](file)
+    transitionTimeTs[idx] = readVal(file, int32)
 
   # Indices into ttinfo structs indicating the changes
   # to be made at the corresponding DST transition.
   var ttInfoIndices = newSeq[byte](tzh_timecnt)
   for idx in 0 ..< tzh_timecnt:
-    ttInfoIndices[idx] = readVal[byte](file)
+    ttInfoIndices[idx] = readVal(file, byte)
 
   # ttinfos which give info on DST transitions.
   var tempTTInfos = newSeq[TempTTInfo](tzh_typecnt)
@@ -146,56 +166,56 @@ proc loadTzInternal(path, dir: string,
     tempTTInfos[idx] = readTempTTInfo(file)
 
   # The array of time zone abbreviation characters.
-  var tzAbbrevChars = readArray[char](file, tzh_charcnt)
+  var tzAbbrevChars = readArray(file, char, tzh_charcnt)
 
   var leapSeconds = newSeq[LeapSecond](tzh_leapcnt)
   for idx in 0 ..< tzh_leapcnt:
-    let time = readVal[int32](file)
-    let total = readVal[int32](file)
+    let time = readVal(file, int32)
+    let total = readVal(file, int32)
     leapSeconds[idx] = LeapSecond(time: time, total: total)
 
   # Indicate whether each corresponding DST transition were specified
   # in standard time or wall clock time.
   var transitionIsStd = newSeq[bool](tzh_ttisstdcnt)
   for idx in 0 ..< tzh_ttisstdcnt:
-      transitionIsStd[idx] = readVal[bool](file)
+      transitionIsStd[idx] = readVal(file, bool)
 
   # Indicate whether each corresponding DST transition associated with
   # local time types are specified in UTC or local time.
   var transitionInUTC = newSeq[bool](tzh_ttisgmtcnt)
   for idx in 0 ..< tzh_ttisgmtcnt:
-    transitionInUTC[idx] = readVal[bool](file)
+    transitionInUTC[idx] = readVal(file, bool)
 
   tzAssert not endOfFile(file)
 
   if version in {'2', '3'}:
     tzAssert readStr(file, 4) == "TZif"
-    tzAssert readVal[char](file) in {'2', '3'}
-    tzAssert readArray[byte](file, 15) == Zeros15
+    tzAssert readVal(file, char) in {'2', '3'}
+    tzAssert readArray(file, byte, 15) == Zeros15
 
     # The number of UTC/local indicators stored in the file.
-    let tzh_ttisgmtcnt = readVal[int32](file)
+    let tzh_ttisgmtcnt = readVal(file, int32)
     # The number of standard/wall indicators stored in the file.
-    let tzh_ttisstdcnt = readVal[int32](file)
+    let tzh_ttisstdcnt = readVal(file, int32)
     # The number of leap seconds for which data is stored in the file.
-    let tzh_leapcnt = readVal[int32](file)
+    let tzh_leapcnt = readVal(file, int32)
     # The number of "transition times" for which data is stored in the file.
-    let tzh_timecnt = readVal[int32](file)
+    let tzh_timecnt = readVal(file, int32)
     # The number of "local time types" for which data is stored in the file(must not be zero).
-    let tzh_typecnt = readVal[int32](file)
+    let tzh_typecnt = readVal(file, int32)
     # The number of characters of "timezone abbreviation strings" stored in the file.
-    let tzh_charcnt = readVal[int32](file)
+    let tzh_charcnt = readVal(file, int32)
 
     # time_ts where DST transitions occur.
     transitionTimeTs = newSeq[int64](tzh_timecnt)
     for idx in 0 ..< tzh_timecnt:
-      transitionTimeTs[idx] = readVal[int64](file)
+      transitionTimeTs[idx] = readVal(file, int64)
 
     # Indices into ttinfo structs indicating the changes
     # to be made at the corresponding DST transition.
     ttInfoIndices = newSeq[byte](tzh_timecnt)
     for idx in 0 ..< tzh_timecnt:
-      ttInfoIndices[idx] = readVal[byte](file)
+      ttInfoIndices[idx] = readVal(file, byte)
 
     # ttinfos which give info on DST transitions.
     tempTTInfos = newSeq[TempTTInfo](tzh_typecnt)
@@ -203,25 +223,25 @@ proc loadTzInternal(path, dir: string,
       tempTTInfos[idx] = readTempTTInfo(file)
 
     # The array of time zone abbreviation characters.
-    tzAbbrevChars = readArray[char](file, tzh_charcnt)
+    tzAbbrevChars = readArray(file, char, tzh_charcnt)
 
     leapSeconds = newSeq[LeapSecond](tzh_leapcnt)
     for idx in 0 ..< tzh_leapcnt:
-      let time = readVal[int64](file)
-      let total = readVal[int32](file)
+      let time = readVal(file, int64)
+      let total = readVal(file, int32)
       leapSeconds[idx] = LeapSecond(time: time, total: total)
 
     # Indicate whether each corresponding DST transition were specified
     # in standard time or wall clock time.
     transitionIsStd = newSeq[bool](tzh_ttisstdcnt)
     for idx in 0 ..< tzh_ttisstdcnt:
-        transitionIsStd[idx] = readVal[bool](file)
+        transitionIsStd[idx] = readVal(file, bool)
 
     # Indicate whether each corresponding DST transition associated with
     # local time types are specified in UTC or local time.
     transitionInUTC = newSeq[bool](tzh_ttisgmtcnt)
     for idx in 0 ..< tzh_ttisgmtcnt:
-      transitionInUTC[idx] = readVal[bool](file)
+      transitionInUTC[idx] = readVal(file, bool)
 
   tzAssert readLine(file).strip() == ""
   # "After  the  second  header and data comes a newline-enclosed,
@@ -295,42 +315,46 @@ proc loadTzInternal(path, dir: string,
   result.name = path
 
   if loadLocation:
-    let opt = loadLocation(path, dir)
+    let opt = loadLocation(dir, path)
     result.location = opt.map(value => value[0])
     result.countries =
       opt.map(value => value[1]).get(@[])
         .filterIt(it.len == 2) # Wrong country codes are silently ignored
         .mapIt(cc(it))
 
-when defined(nimdoc):
-  # TODO: Should I respect TZDIR env variable?
-  const PosixTimezonesDir* = "" ## Doc comment here
-elif defined(PosixTimezonesDir):
-  const PosixTimezonesDir* {.strdefine.} = ""
-elif defined(posix):
-  const PosixTimezonesDir* = "/usr/share/zoneinfo/"
-else:
-  const PosixTimezonesDir* = ""
+proc resolvePosixTzPath(path: string): (string, string) =
+  if path.isAbsolute:
+    result = ("", path)
+  else:
+    let dir = getEnv("TZDIR", "/usr/share/zoneinfo/")
+    result = (dir, path)
 
-proc loadTz*(path: string, dir = PosixTimezonesDir): Timezone =
+proc loadPosixTz*(path: string): Timezone
+    {.raises: [IOError, TzFileParsingError].}=
   ## Load a timezone from a posix timezone file.
   ##
-  ## If ``path`` is a relative path, it's treated as relative to ``dir``.
-  ## The timezone will use ``path`` as name (typically this will result
-  ## in the name matching the IANA timezone name).
-  let tzInternal = loadTzInternal(path, dir)
+  ## If `path` is relative it's interpreted as relative to the system
+  ## timezone dir, meaning that e.g `loadPosixTz"Europe/Stockholm"` works.
+  ## The timezone dir is `"/usr/share/zoneinfo/"` by default, but can be
+  ## overriden by setting the `TZDIR` environment variable.
+  let (dir, path) = resolvePosixTzPath(path)
+  let tzInternal = loadTzInternal(dir, path)
   result = newTimezone(tzInternal)
 
-proc loadTzInfo*(path: string, dir = PosixTimezonesDir): TimezoneInfo =
+proc loadPosixTzInfo*(path: string): TimezoneInfo
+    {.raises: [IOError, TzFileParsingError].}=
   ## Load a timezone and related metadata from a posix timezone file.
   ##
-  ## The metadata is loaded from the file ``zone1970.tab``, which is expected
-  ## to exists in ``dir`` (this will be true on a typicall posix system).
+  ## The metadata is loaded from the file `zone1970.tab`, which is expected
+  ## to exist in the system timezone dir.
   ##
-  ## If ``path`` is a relative path, it's treated as relative to ``dir``.
-  ## The timezone will use ``path`` as name (typically this will result
-  ## in the name matching the IANA timezone name).
-  let tzInternal = loadTzInternal(path, dir, loadLocation = true)
+  ## The `path` argument must be a relative path, and will be interpreted as
+  ## relative to the system timezone dir.
+  ## The timezone dir is `"/usr/share/zoneinfo/"` by default, but can be
+  ## overriden by setting the `TZDIR` environment variable.
+  doAssert not path.isAbsolute, "`path` must be relative: " & path
+  let (dir, path) = resolvePosixTzPath(path)
+  let tzInternal = loadTzInternal(dir, path, loadLocation = true)
   result.timezone = newTimezone(tzInternal)
   result.location = tzInternal.location
   result.countries = tzInternal.countries.mapIt($it)
@@ -346,15 +370,25 @@ iterator walkDirRecRelative(dir: string): string =
       if k in {pcFile}:
         yield rel
 
-proc loadTzDb*(filter: seq[string],
-               dir = PosixTimezonesDir): timezones.TimezoneDb =
-  ## Load all available timezones from ``dir``.
+proc loadPosixTzDb*(dir = ""):
+    timezones.TimezoneDb {.raises: [IOError, TzFileParsingError].} =
+  ## Load all available timezones and metadata from ``dir``.
+  ##
+  ## If `dir` is the empty string, the system timezone dir will be used.
+  ##
+  ## The version field of the returned ``TimezoneDb`` will be set to the
+  ## empty string, because there's no reliable way extract the version from a
+  ## timezone dir.
+  let dir =
+    if dir == "":
+      let (dir, _) = resolvePosixTzPath("")
+      dir
+    else:
+      dir
   var zones = newSeq[TimezoneInternal]()
-  for tzName in walkDirRecRelative(dir):
-    if tzName.splitFile.ext != "" or tzName in ["leapseconds", "+VERSION"]:
-      continue
-    if filter.len > 0 and tzname notin filter:
+  for path in walkDirRecRelative(dir):
+    if path.splitFile.ext != "" or path in ["leapseconds", "+VERSION"]:
       continue
     # TODO: Don't use loadLocation here!
-    zones.add(loadTzInternal(tzname, dir, loadLocation = true))
+    zones.add(loadTzInternal(dir, path, loadLocation = true))
   result = timezones.TimezoneDb(initTimezoneDb("", zones))